% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_meta_ann.R
\name{ann_metaf}
\alias{ann_metaf}
\title{ann_metaf}
\usage{
ann_metaf(catchment, data_inputs, bs, epochs, ensemble_runs, train_mean,
  train_sd, x_train, y_train, x_val, y_val, x_test, y_test, n_features,
  model_name, user_name)
}
\arguments{
\item{ensemble_runs}{}
}
\value{

}
\description{
ann_metaf
}
