% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitter_for_lstm.R
\name{data_splitter_for_lstm}
\alias{data_splitter_for_lstm}
\title{data_splitter_for_lstm}
\usage{
data_splitter_for_lstm(data, data_name)
}
\arguments{
\item{data}{Time series data as a data frame with the first column containing a POSIXct data and is named "date".}
}
\value{

}
\description{
data_splitter_for_lstm
}
