% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_meta_lstm.R
\name{lstm_metaf}
\alias{lstm_metaf}
\title{lstm_metaf}
\usage{
lstm_metaf(x_train, y_train, x_val, y_val, data_inputs, train_mean,
  train_sd, ts, u, bs, epochs, LSTM_type, n_predictions, n_features)
}
\arguments{
\item{n_features}{}
}
\value{

}
\description{
LSTM modelling function. Not available to the user. Use wt_lstm.
}
