% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_nse.R
\name{rmse_nse}
\alias{rmse_nse}
\title{rmse_nse}
\usage{
rmse_nse(model, val)
}
\arguments{
\item{model}{trained model}

\item{val}{validation data}
}
\value{
data frame with RMSE and NSE
}
\description{
rmse_nse
}
