% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_catchment}
\alias{test_catchment}
\title{Stream water temperature, discharge and meteorological data of the wateRtemp test catchment.}
\format{
A data frame with 6851 rows and 10 variables:
\describe{
\item{year}{Years as integer.}
\item{month}{Months as integer.}
\item{day}{Days as integer.}
\item{Q}{Daily mean discharge in m³/s at the gauging station of the test catchment.}
\item{P}{Daily precipitation sums in mm, averaged over the whole catchment.}
\item{Ta_min}{Daily minimum air temperature °C, averaged over the whole catchment.}
\item{Ta_max}{Daily maximum air temperature °C, averaged over the whole catchment.}
\item{Ta}{Daily mean air temperature °C, averaged over the whole catchment.}
\item{wt}{Daily mean stream water temperature in °C at the gauging station of the test catchment.}
\item{GL}{Daily mean global radiation in W/m², averaged over the whole catchment.}
}
}
\source{
Feigl, M., Lebidzinski, K., Herrnegger, M. and Schulz, K.: Machine learning methods for stream water temperature prediction
}
\usage{
data(test_catchment)
}
\description{
A synthetic dataset for mean daily stream water temperature prediction. This data sets is made for testing all wateRtemp functionalities and has characteristics mimicking real world data.
}
\examples{
data(test_catchment)
summary(test_catchment)
}
\references{
Feigl, M., Lebidzinski, K., Herrnegger, M. and Schulz, K.: Machine learning methods for stream water temperature prediction
}
\keyword{datasets}
