% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Traisen}
\alias{Traisen}
\title{Stream water temperature, discharge and meteorological data of the Traisen catchment.}
\format{
A data frame with 6471 rows and 10 variables:
\describe{
\item{year}{Years as integer.}
\item{month}{Months as integer.}
\item{day}{Days as integer.}
\item{Q}{Daily mean discharge in m³/s at the Windpassing gauging station of the Traisen catchment.}
\item{P}{Daily precipitation sums in mm, averaged over the whole catchment.}
\item{Ta_min}{Daily minimum air temperature °C, averaged over the whole catchment.}
\item{Ta_max}{Daily maximum air temperature °C, averaged over the whole catchment.}
\item{Ta}{Daily mean air temperature °C, averaged over the whole catchment.}
\item{wt}{Daily mean stream water temperature in °C at the Windpassing gauging station of the Traisen catchment.}
\item{GL}{Daily mean global radiation in W/m², averaged over the whole catchment.}
}
}
\source{
Feigl, M., Lebidzinski, K., Herrnegger, M. and Schulz, K.: Machine learning methods for stream water temperature prediction
}
\usage{
data(Traisen)
}
\description{
A dataset for mean daily stream water temperature prediction.
}
\examples{
data(Traisen)
summary(Traisen)
}
\references{
Feigl, M., Lebidzinski, K., Herrnegger, M. and Schulz, K.: Machine learning methods for stream water temperature prediction
}
\keyword{datasets}
