% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_ann.R
\name{wt_ann}
\alias{wt_ann}
\title{wt_ann}
\usage{
wt_ann(catchment, data_inputs = NULL, model_or_optim, bs = c(10, 50,
  100), epochs = c(100), ensemble_runs = 100)
}
\arguments{
\item{ensemble_runs}{}
}
\description{
wt_ann
}
