% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_ann_DataXGB.R
\name{wt_ann_DataXGB}
\alias{wt_ann_DataXGB}
\title{wt_ann}
\usage{
wt_ann_DataXGB(
  catchment,
  data_inputs = NULL,
  model_or_optim,
  bs = c(10, 50, 100),
  epochs = c(100),
  ensemble_runs = 100,
  user_name = "R2D2"
)
}
\arguments{
\item{ensemble_runs}{}
}
\value{

}
\description{
wt_ann
}
\details{
if ensemble_runs >= 100 only the best 10% are used for the prediction, if ensemble runs < 100, only the best 10 are used. In case there are less than 10 ensemble runs, all are used for prediction.
}
