% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_fnn.R
\name{wt_fnn}
\alias{wt_fnn}
\title{wt_fnn}
\usage{
wt_fnn(
  train_data,
  test_data = NULL,
  catchment = NULL,
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL,
  n_iter = 40,
  n_random_initial_points = 20,
  epochs = 100,
  early_stopping_patience = 5,
  ensemble_runs = 5,
  bounds_layers = c(1, 5),
  bounds_units = c(5, 200),
  bounds_dropout = c(0, 0.2),
  bounds_batch_size = c(5, 150),
  initial_grid_from_model_scores = TRUE
)
}
\arguments{
\item{initial_grid_from_model_scores}{}
}
\value{

}
\description{
wt_fnn
}
