% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_fnn_optimizer.R
\name{wt_fnn_optimizer}
\alias{wt_fnn_optimizer}
\title{wt_fnn_optimizer}
\usage{
wt_fnn_optimizer(
  catchment,
  data_inputs = NULL,
  n_iter = 20,
  n_random_initial_points = 20,
  epochs = 100,
  early_stopping_patience = 5,
  ensemble_runs = 3,
  bounds_layers = c(1, 5),
  bounds_units = c(5, 200),
  bounds_dropout = c(0, 0.2),
  bounds_batch_size = c(2, 150),
  top_n_models = 3,
  initial_grid_from_model_scores = TRUE,
  user_name = "R2D2"
)
}
\arguments{
\item{user_name}{}
}
\value{

}
\description{
wt_fnn_optimizer
}
