% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_lm.R
\name{wt_lm}
\alias{wt_lm}
\title{wt_lm}
\usage{
wt_lm(
  train_data,
  test_data = NULL,
  catchment = NULL,
  type = "stepLM",
  cv_mode = "repCV",
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL
)
}
\arguments{
\item{train_data}{Data frame containing training data created by using wt_preprocessing()}

\item{test_data}{Data frame containing test data created by using wt_preprocessing()}

\item{catchment}{Catchment name as string, used for storing results in current working directory.}

\item{type}{Can be either "LM" for a multiple regression model of the form wt ~ Ta + Q, or "stepLM" for a step-wise linear regression model using all available variables and their interactions.}

\item{cv_mode}{Cross-validation mode. Only relvenat when using type="stepLM". Can either be "repCV" for a 5times repeated 10-fold CV or "timeseriesCV" for a timeslice CV using intial window=730, horizon=90 and skip=60.}

\item{model_name}{Name of this particular model run as string, used for storing results in the catchment folder.}

\item{no_cores}{Number of cores used for computation. If NULL parallel::detectCores() - 1 is applied.}

\item{seed}{Random seed.}
}
\description{
Multiple linear regression and step-wise linear regression implementation for stream water temperature prediction including Bayesian hyperparameter optimization. All results are stored automatically in the folder catchment/model_name.
}
\examples{
\dontrun{
data(Aschach)
wt_preprocess(Aschach)
train_data <- feather::read_feather("Aschach/train_data.feather")
test_data <- feather::read_feather("Aschach/test_data.feather")

wt_lm(train_data, test_data, "Aschach", "stepLM", "repCV", "standard_stepLM")
}
}
