% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_lstm.R
\name{wt_lstm}
\alias{wt_lstm}
\title{wt_lstm}
\usage{
wt_lstm(catchment, data_inputs = NULL, ts = c(30, 90, 180, 365),
  u = c(30, 60, 90), bs = c(10, 50, 100), epochs = c(100),
  LSTM_type = NULL, n_predictions = c(1, 3, 7))
}
\arguments{
\item{n_predictions}{}
}
\value{
None
}
\description{
wt_lstm
}
