% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_preprocessing.R
\name{wt_preprocess}
\alias{wt_preprocess}
\title{wt_preprocess}
\usage{
wt_preprocess(data, year_range, catchment)
}
\arguments{
\item{catchment}{}
}
\value{

}
\description{
wt_preprocess
}
