% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_randomforest.R
\name{wt_randomforest}
\alias{wt_randomforest}
\title{wt_randomforest}
\usage{
wt_randomforest(
  train_data,
  test_data = NULL,
  catchment = NULL,
  cv_mode = "repCV",
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL
)
}
\arguments{
\item{train_data}{Data frame containing training data created by using wt_preprocessing()}

\item{test_data}{Data frame containing test data created by using wt_preprocessing()}

\item{catchment}{Catchment name as string, used for storing results in current working directory.}

\item{cv_mode}{Cross-validation mode. Can either be "repCV" for a 5times repeated 10-fold CV or "timeseriesCV" for a timeslice CV using intial window=730, horizon=90 and skip=60.}

\item{model_name}{Name of this particular model run as string, used for storing results in the catchment folder.}

\item{no_cores}{Number of cores used for computation. If NULL parallel::detectCores() - 1 is applied.}

\item{seed}{Random seed.}
}
\description{
Randomforest implementation for stream water temperature prediction including Bayesian hyperparameter optimization. All results are stored automatically in the folder catchment/model_name.
}
\examples{
\dontrun{
data(test_catchment)
wt_preprocess(test_catchment)
train_data <- feather::read_feather("test_catchment/train_data.feather")
test_data <- feather::read_feather("test_catchment/test_data.feather")

wt_randomforest(train_data, test_data, "test_catchment", "repCV", "standard_rf")
}
}
