% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_result_analysis.R
\name{wt_result_analysis}
\alias{wt_result_analysis}
\title{wt_result_analysis}
\usage{
wt_result_analysis(catchment, model, folder_name, epochs, bs, plot)
}
\arguments{
\item{plot}{}
}
\value{
None
}
\description{
Plots diagnostic plots for analysing the differen model results.
}
