% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_result_analysis2.R
\name{wt_result_analysis2}
\alias{wt_result_analysis2}
\title{wt_result_analysis}
\usage{
wt_result_analysis2(catchment, model, folder1, folder2, years = "2015")
}
\arguments{
\item{plot}{}
}
\value{
None
}
\description{
Plots diagnostic plots for analysing the differen model results.
}
