% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_rnn_fmon.R
\name{wt_rnn_fmon}
\alias{wt_rnn_fmon}
\title{wt_rnn}
\usage{
wt_rnn_fmon(
  catchment,
  data_inputs = NULL,
  rnn_type = NULL,
  layers = NULL,
  n_predictions = c(1, 3, 7),
  timesteps = c(60, 90),
  units = c(30, 60),
  dropout = 0,
  batch_size = c(10, 50, 100),
  epochs = 100,
  early_stopping_patience = 5,
  ensemble_runs = 5,
  user_name = "R2D2"
)
}
\arguments{
\item{catchment}{Name of the folder with the data given as a string.}

\item{data_inputs}{The kind of data to be used for the model run: "simple", "precip", "radiation", "all". See Details}

\item{layers}{Number RNN LSTM layers to use, integer}

\item{n_predictions}{Number of timesteps to be predicted by LSTM, integer}

\item{epochs}{Number of epochs, integer}

\item{ts}{Number of timesteps used for prediction, integer}

\item{u}{Number of units of the LSTM layers, integer}

\item{bs}{batch size for training, integer}
}
\value{
None
}
\description{
wt_rnn
}
