% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_rnn_optimizer.R
\name{wt_rnn_optimizer}
\alias{wt_rnn_optimizer}
\title{wt_rnn_optimizer}
\usage{
wt_rnn_optimizer(
  catchment,
  data_inputs = NULL,
  rnn_type = NULL,
  n_iter = 20,
  n_random_initial_points = 20,
  n_predictions = 1,
  epochs = 100,
  early_stopping_patience = 5,
  ensemble_runs = 3,
  bounds_layers = c(1, 4),
  bounds_timesteps = c(10, 150),
  bounds_units = c(5, 150),
  bounds_dropout = c(0, 0.5),
  bounds_batch_size = c(5, 150),
  top_n_models = 3,
  initial_grid_from_model_scores = FALSE,
  user_name = "R2D2"
)
}
\arguments{
\item{user_name}{}
}
\value{

}
\description{
wt_rnn_optimizer
}
