% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_single_rnn_fmon.R
\name{wt_single_rnn_fmon}
\alias{wt_single_rnn_fmon}
\title{single_rnn}
\usage{
wt_single_rnn_fmon(
  catchment,
  x_train,
  y_train,
  x_val,
  y_val,
  x_test,
  y_test,
  test,
  rnn_test,
  user_name,
  data_inputs,
  rnn_type,
  timesteps,
  units,
  dropout,
  layers,
  batch_size,
  epochs,
  early_stopping_patience,
  n_predictions,
  n_features,
  ensemble_runs,
  return_flag
)
}
\arguments{
\item{catchment}{Name of the folder with the data given as a string.}

\item{n_features}{}
}
\value{

}
\description{
RNN modelling function. Not available to the user. Use wt_rnn.
}
