% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_xgboost.R
\name{wt_xgboost}
\alias{wt_xgboost}
\title{wt_xgboost}
\usage{
wt_xgboost(
  train_data,
  test_data = NULL,
  catchment = NULL,
  cv_mode = "repCV",
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL,
  n_iter = 40,
  n_random_initial_points = 20
)
}
\arguments{
\item{n_random_initial_points}{}
}
\value{

}
\description{
wt_xgboost
}
