% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_xgboost_simple.R
\name{wt_xgboost_simple}
\alias{wt_xgboost_simple}
\title{wt_xgboost_simple}
\usage{
wt_xgboost_simple(
  catchment,
  data_inputs = NULL,
  model_or_optim,
  cv_mode,
  parameter_list,
  plot_ts = FALSE,
  save_importance_plot = FALSE,
  user_name = "R2D2"
)
}
\arguments{
\item{catchment}{Name of the folder with the data as a string.}

\item{data_inputs}{The kind of data to be used for the model run: "simple", "precip", "radiation", "all". See Details}

\item{model_or_optim}{"model" if a pretrained model should be load, or "optim" if a new model should be trained}

\item{cv_mode}{The crossvalidation scheme to be used. Can be either: "timeslice", or "repCV"}

\item{plot_ts}{Should a dygraphy plot of the prediction and observation be plotted? TRUE/FALSE}

\item{save_importance_plot}{Should the importance plot for XGBoost be saved in the folder. TRUE/FALSE}

\item{n_iter}{Number of iterations for the bayesian hyperparameter optimization as a integer.}
}
\value{
None
}
\description{
XGBoost modelling for water temperature data prepared with wt_preprocessing. Can either be used for loading pre-trained models or train a new model. The training is done in parallel and can take a while. yperparameter optimization is done using grid search and the rBayesianOptimization packages.
}
\details{
This function is able to train or load a pre-trained model for water temperature prediction. The features used for the prediction are defined by the "data_inputs" variable. "simple" means only discharge Q and mean air temperature Tmean are used. "precip" means that additional to "simple", precipitation data "RR" is used. "radiation" means that additional to "simple", longwave radiation data "GL" is used. "all" means that "simple" + precipitation + radiation data is used.
}
