% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_reshaper.R
\name{x_reshaper}
\alias{x_reshaper}
\title{x_reshaper}
\usage{
x_reshaper(x, n_timesteps, n_predictions)
}
\arguments{
\item{x}{matrix with features for the LSTM}

\item{n_timesteps}{number of timesteps used for prediction}

\item{n_predictions}{number of timesteps to be predicted by LSTM}
}
\value{
returns an 3D array with (number observations, timesteps, number of features)
}
\description{
Function for reshaping data for LSTM applications. x_reshaper reshapes the feature data to a 3D array of features with (number observations, timesteps, number of features). y_reshaper reshapes the the data to a 1D array with labels corresponding to the 3D feature array.
}
