% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_reshaper.R
\name{y_reshaper}
\alias{y_reshaper}
\title{y_reshaper}
\usage{
y_reshaper(y, n_timesteps, n_predictions)
}
\arguments{
\item{y}{matrix with labels for the LSTM}

\item{n_timesteps}{number of timesteps used for prediction}

\item{n_predictions}{number of timesteps to be predicted by LSTM}
}
\value{
returns an 1D array with (number observations)
}
\description{
Function for reshaping data for LSTM applications. x_reshaper reshapes the label data to a 3D array of features with (number observations, timesteps, number of features). y_reshaper reshapes the the data to a 1D array with labels corresponding to the 3D feature array.
}
