% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_fnn.R
\name{wt_fnn}
\alias{wt_fnn}
\title{wt_fnn}
\usage{
wt_fnn(
  catchment,
  data_inputs = NULL,
  batch_size = c(10, 50, 100),
  layers = c(2, 4, 8, 15, 20),
  units = c(10, 30, 50, 100),
  dropout = 0,
  epochs = c(100),
  early_stopping_patience = 5,
  ensemble_runs = 100,
  user_name = "R2D2"
)
}
\arguments{
\item{user_name}{}
}
\value{

}
\description{
wt_fnn
}
