% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_lm.R
\name{wt_lm}
\alias{wt_lm}
\title{wt_lm}
\usage{
wt_lm(catchment, data_inputs = NULL, type = NULL, user_name = "R2D2")
}
\arguments{
\item{catchment}{Name of the folder containing the data. Given as a string}

\item{type}{}
}
\description{
wt_lm
}
