% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_lm.R
\name{wt_lm}
\alias{wt_lm}
\title{wt_lm}
\usage{
wt_lm(
  train_data,
  test_data = NULL,
  catchment = NULL,
  type = "stepLM",
  cv_mode = "repCV",
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
wt_lm
}
