% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_lstm.R
\name{wt_lstm}
\alias{wt_lstm}
\title{wt_lstm}
\usage{
wt_lstm(catchment, data_inputs = NULL, ts = c(30, 90, 180, 365),
  u = c(30, 60, 90), bs = c(10, 50, 100), epochs = c(100),
  lstm_layers = NULL, n_predictions = c(1, 3, 7))
}
\arguments{
\item{catchment}{Name of the folder with the data given as a string.}

\item{data_inputs}{The kind of data to be used for the model run: "simple", "precip", "radiation", "all". See Details}

\item{ts}{Number of timesteps used for prediction, integer}

\item{u}{Number of units of the LSTM layers, integer}

\item{bs}{batch size for training, integer}

\item{epochs}{Number of epochs, integer}

\item{lstm_layers}{Number of LSTM layers to use, integer}

\item{n_predictions}{Number of timesteps to be predicted by LSTM, integer}
}
\value{
None
}
\description{
wt_lstm
}
