% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_randomforest.R
\name{wt_randomforest}
\alias{wt_randomforest}
\title{wt_randomforest}
\usage{
wt_randomforest(
  train_data,
  test_data = NULL,
  catchment = NULL,
  cv_mode = "repCV",
  model_name = NULL,
  no_cores = parallel::detectCores() - 1,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
wt_randomforest
}
