% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_randomforest.R
\name{wt_randomforest}
\alias{wt_randomforest}
\title{wt_RandomForest}
\usage{
wt_randomforest(catchment, data_inputs = NULL, model_or_optim, cv_mode,
  no_cores = detectCores() - 1, plot_ts = FALSE,
  save_importance_plot = FALSE)
}
\arguments{
\item{catchment}{Name of the folder with the data as a string.}

\item{model_or_optim}{"model" if a pretrained model should be load, or "optim" if a new model should be trained}

\item{cv_mode}{The crossvalidation scheme to be used. Can be either: "timeslice", or "repCV"}

\item{no_cores}{Number of cores used for training as integer.}

\item{plot_ts}{Should a dygraphy plot of the prediction and observation be plotted? TRUE/FALSE}

\item{save_importance_plot}{Should the importance plot for the random forest be saved in the folder. TRUE/FALSE}
}
\value{
None, results are saved in RF folder which is created inside the catchment folder.
}
\description{
Random Forest modelling for water temperature data prepared with wt_preprocessing. Can either be used for loading pre-trained models or train a new model. The training is done in parallel and can take a while. Hyperparameter optimization is done by grid search.
}
\examples{
wt_preprocess("Ybbs", data = data, year_range = c(1981, 2015))
wt_RandomForest(catchment = "Ybbs",
                model_or_optim = "optim",
                cv_mode = "timeslice",
                no.cores = 17,
                plot_ts = FALSE,
                save_importance_plot = TRUE)

}
